BoardConfig: Enable watchdog_v2 IPI pings
Since the -ng kernel branch used by suzu uses IPI to keep
the device functioning in case of GPU IOMMU hangs, it might be
helpful to additionally check for other CPUs being online.

The rather advanced power saving mechanisms of suzu also cause
unnecessary CPUs to be offlined as the kernel sees fit,
which is taken into account for this setting.